requirements.txt
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-dotenv==1.0.0
typer[all]>=0.9.0
tqdm>=4.66.1
rich>=13.7.0
playwright>=1.40.0
html2text>=2020.1.16
pytest>=7.4.3
openpyxl>=3.1.2
